requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.7
azure-identity>=1.12.0
azure-keyvault-keys>=4.5.0
cryptography==41.0
gunicorn==20.1.0
